Refactor bmp2chr workflow
PatternTable is now the base type of an image.  It is made up of a list
of tiles.  It's not fully fleshed out yet, but it will implement the
Image interface from the standard library (and maybe even the Drawer
interface).

The bmp2chr has had it's feature set reduced until a bunch of the
features can be re-implemented (mostly, anything passed from the
command line).