Implement writing tile IDs to a file
This change adds the option --write-ids <filename>.  It can only be used
with --remove-duplicates and --remove-empty.

With this option all tile IDs will be listed in the order they appear in
the input image.  They will use the new IDs that have been assigned
after tile reduction takes place.

The output file will have either two or four unnamed labels.  If the
image spans across two nametables there will be four labels, two
otherwise.  For each pair of labels, the first is for the number of
tiles in the data and the second is the tile data.