Refactor code into Read/Write/Import/Export functions
This change makes a lot of internal things private that don't need to be
exposed publicly.  The main method of interacting with the library is
through four new functions:

Read() and Import() read a `.studybox` file, and a `.json` file (along
with it's associated directory), respectively.

Write() and Export() write a `.studybox` file, and a `.json` file in
addition to all the parts of the tape unpacked to a directory,
respectively.