Add fontutil command
This is a short command to handle bitmap fonts.  It takes a single
bitmap image of a font with the runes at their ASCII value tile offset
in the image.  Three files will be written: an ASM file containing the
first bitplane of each tile, a file containing .charmap directives
remapping ascii to the font, and a file with a list of widths for each
character.  All three filenames need to be passed as options on the
command line.