Add text2chr utility
This utility will take a string of text and a font image and turn that
into a CHR file with variable width text.  It'll also write a metadata
file that can be used to draw the text.

The code for tihs is an absolute mess, but it "works".  Some of the
coordinate stuff doesn't seem correct.  There's still a bunch of
commented out debug code to figure some of that out.