[metatiles] Fix tile ID calculation with 2+ rows
Fixed the tile ID calculation when there is more than a single row of
metatiles.  The original formula had ImageMetatileWidth instead of
ImageTileWidth in the first block of multiplication, so instead of using
an image with of 16 8x8 pixel tiles, 8 8x8 pixel tiles were used.  This
essentially ended up starting the second metatile row with the bottom
half of the first metatile row when using the default metatile size of
2x2.

Given the following for each 8x8 pixel tile in a metatile:
    Y = Current metatile row
    X = Current metatile column
    I = Current tile row inside metatile
    J = Current tile column inside metatile
    ImageTileWidth = Width of source image in 8x8 pixel tiles
    ImageMetatileWidth = Width of source image in metatiles

ID = (Y * MetaTileHeight * ImageTileWidth) +
     (X * MetaTileWidth) +
     (I * MetaTileWidth * ImageMetatileWidth) +
     J